refactor(process): migrate dashboard module and restructure files
