+++ src/modules/process/ui/screens/DashboardScreen.tsx
@@ -2,7 +2,7 @@ import { A, useNavigate } from '@solidjs/router'
 import type { JSX } from 'solid-js'
 import { createResource, createSignal, For, Show } from 'solid-js'
 import z from 'zod'
-import { presentProcessList } from '~/modules/dashboard/application/processListPresenter'
+import { presentProcessList } from '~/modules/process/application/queries/processList.presenter'
 import type { CreateProcessInput } from '~/modules/process/interface/http/process.schemas'
 import type { CreateProcessDialogFormData } from '~/modules/process/ui/CreateProcessDialog'
 import { CreateProcessDialog } from '~/modules/process/ui/CreateProcessDialog'
